pyproject.toml
[tool.poetry]
name = "fastapi_test_task"
version = "0.1.0"
description = "Fast API test task"
authors = ["Artemooon <artem.logachov773@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
tortoise-orm = "^0.17.7"
asyncpg = "^0.24.0"
aerich = "^0.5.8"
pydantic = {extras = ["email"], version = "^1.8.2"}
passlib = "^1.7.4"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
